refactor - note that it's not worth to split it off into a crate…
…even though it will have additional dependencies.

However, it's so tightly coupled with criner that without even more
splitting work, it's not benefit to do.

Besides, who needs crates on crates.io that are only useful to one
other program?

What about 'anonymous' crates, those that are part of another
crate but not directly available on crates.io?