Skip all recently yanked versions when reporting…
…better than before, but I am still unsure about handling of yanked
crates. At least you can't add a yanked crate as a dependency, but
it could still be dependent on by various crates and thus increasing
download volume.

However, these would never be a problem for people adding the crate,
which is why ignoring recently yanked crates is the right thing